Make gateways horizontally safe: mesh-wide sessions, reconnect tokens, generation fencing, draining, load reports, peer authentication
Session ids are 64-bit and unique across gateways (incarnation << 32 | sequence); the worker keys player state by them.
Every welcomed client gets a signed session token and can reconnect through any gateway of the mesh with it: the
gateway reclaims the session with a newer connection generation, the worker re-announces the pawn and fences the
old gateway (stale despawns and inputs are ignored), and pawns whose link dropped are kept for SessionReclaimSeconds.
Gateways report load on the control-plane heartbeat (GatewayStats), can be asked to drain (POST /api/gateways/drain),
and authenticate workers and other gateways with an HMAC credential derived from the mesh token. Protocol version 13.

+++ Packages/com.1by3.nebula/Runtime/Core/NebulaGameMode.cs
@@ -12,7 +12,7 @@ namespace Nebula
         /// Create and spawn the player's entity inside <paramref name="container"/>. Return the spawned identity.
         /// Override this or the <see cref="PlayerInfo"/> overload (which also gets the player's identity across sessions).
         /// </summary>
-        public virtual NetworkIdentity OnSpawnPlayer(NebulaWorker worker, uint clientId, string playerName, Container container)
+        public virtual NetworkIdentity OnSpawnPlayer(NebulaWorker worker, ulong clientId, string playerName, Container container)
         {
             NebulaLog.Error($"{GetType().Name} overrides neither OnSpawnPlayer overload; client {clientId} gets no pawn");
             return null;
@@ -35,8 +35,11 @@ namespace Nebula
     /// <summary>Who is joining, as the gateway told the worker (<see cref="NebulaGameMode.OnSpawnPlayer(NebulaWorker, in PlayerInfo, Container)"/>).</summary>
     public readonly struct PlayerInfo
     {
-        /// <summary>This session's connection id. Changes every time the player connects.</summary>
-        public readonly uint ClientId;
+        /// <summary>
+        /// The player's session id: unique across every gateway of the mesh, and kept across a reconnection with a
+        /// session token (the player continues with the same pawn). A new session gets a new id.
+        /// </summary>
+        public readonly ulong ClientId;
         /// <summary>The display name the client asked for.</summary>
         public readonly string Name;
         /// <summary>The player's identity across sessions (<see cref="PlayerIdentity"/>): from the OpenID token they presented, or the anonymous token the gateway issued them.</summary>
@@ -44,7 +47,7 @@ namespace Nebula
         /// <summary>The client is a headless bot.</summary>
         public readonly bool IsBot;
 
-        public PlayerInfo(uint clientId, string name, string identity, bool isBot)
+        public PlayerInfo(ulong clientId, string name, string identity, bool isBot)
         {
             ClientId = clientId;
             Name = name ?? "";
